+++ src/main/java/com/baomidou/mybatisplus/extension/NonValue.java
@@ -1,6 +1,0 @@
-package com.baomidou.mybatisplus.extension;
-
-import java.io.Serializable;
-
-public class NonValue implements Serializable {
-}