+++ src/main/java/com/baomidou/mybatisplus/extension/SingleBooleanValue.java
@@ -1,4 +1,0 @@
-package com.baomidou.mybatisplus.extension;
-
-public class SingleBooleanValue extends SingleValue<Boolean> {
-}