+++ src/main/java/com/baomidou/mybatisplus/extension/value/SingleDateValue.java
@@ -1,0 +1,6 @@
+package com.baomidou.mybatisplus.extension.value;
+
+import java.util.Date;
+
+public class SingleDateValue extends SingleValue<Date> {
+}